Refactor query
