+++ src/Couchbase/Protostellar/Internal/QueryRequestConverter.php
@@ -22,15 +22,14 @@ declare(strict_types=1);
 namespace Couchbase\Protostellar\Internal;
 
 use Couchbase\Exception\InvalidArgumentException;
+use Couchbase\Protostellar\Generated\KV\V1\MutationToken;
 use Couchbase\Protostellar\Generated\Query\V1\QueryRequest\ScanConsistency;
 use Couchbase\QueryProfile;
-use Couchbase\Protostellar\Generated\KV\V1\MutationToken;
-use Couchbase\Protostellar\Generated\Query\V1\QueryResponse\MetaData;
 use Couchbase\QueryScanConsistency;
 use Google\Protobuf\Duration;
 use Couchbase\Protostellar\Generated\Query\V1\QueryRequest;
 
-class QueryConverter
+class QueryRequestConverter
 {
     /**
      * @throws InvalidArgumentException
@@ -82,22 +81,6 @@ class QueryConverter
             $request["profile_mode"] = self::convertQueryProfile($exportedOptions["profile"]);
         }
         return $request;
-    }
-
-    /**
-     * @param array $response
-     * @return array
-     * @internal
-     */
-    public static function convertQueryResult(array $response): array
-    {
-        $finalArray = [];
-        foreach ($response as $result) {
-            $finalArray["rows"][] = SharedUtils::toArray($result->getRows());
-        }
-        $finalArray["rows"] = call_user_func_array('array_merge', $finalArray["rows"]);
-        $finalArray["meta"] = self::convertMetaData(end($response)->getMetaData());
-        return $finalArray;
     }
 
     private static function getTuningOptions(array $exportedOptions): array
@@ -191,48 +174,5 @@ class QueryConverter
             default:
                 throw new InvalidArgumentException("Unexpected query profile");
         }
-    }
-
-    private static function convertMetaData(MetaData $metadata): array
-    {
-        $finalMetaData = [];
-        $finalMetaData["requestId"] = $metadata->getRequestId();
-        $finalMetaData["clientContextId"] = $metadata->getClientContextId();
-        $finalMetaData["signature"] = $metadata->getSignature();
-        if ($metadata->hasMetrics()) {
-            $finalMetaData["metrics"] = self::convertMetrics($metadata->getMetrics());
-        }
-        $finalMetaData["status"] = SharedUtils::convertStatus(MetaData\Status::name($metadata->getStatus()));
-        if ($metadata->getWarnings()->count()) {
-            $finalMetaData["warnings"] = self::convertWarnings(SharedUtils::toArray($metadata->getWarnings()));
-        }
-        if ($metadata->hasProfile()) {
-            $finalMetaData["profile"] = $metadata->getProfile();
-        }
-        return $finalMetaData;
-    }
-
-    private static function convertWarnings(array $warnings): array
-    {
-        $warningsArr = [];
-        foreach ($warnings as $warning) {
-            $warningsArr[] = ["code" => $warning->getCode(), "message" => $warning->getMessage()];
-        }
-        return $warningsArr;
-    }
-
-
-    private static function convertMetrics(MetaData\Metrics $metrics): array
-    {
-        $finalMetrics = [];
-        $finalMetrics["elapsedTime"] = (intval($metrics->getElapsedTime()->getSeconds()) * 1e9) + $metrics->getElapsedTime()->getNanos();
-        $finalMetrics["executionTime"] = (intval($metrics->getExecutionTime()->getSeconds()) * 1e9) + (($metrics->getExecutionTime()->getNanos()));
-        $finalMetrics["resultCount"] = intval($metrics->getResultCount());
-        $finalMetrics["resultSize"] = intval($metrics->getResultSize());
-        $finalMetrics["mutationCount"] = intval($metrics->getMutationCount());
-        $finalMetrics["sortCount"] = intval($metrics->getSortCount());
-        $finalMetrics["errorCount"] = intval($metrics->getErrorCount());
-        $finalMetrics["warningCount"] = intval($metrics->getWarningCount());
-        return $finalMetrics;
     }
 }
